requirements.txt
streamlit
torch
torchvision
scikit-learn
numpy
pillow
joblib
matplotlib
scikit-image
